Forbid recomputing the last message sent
For a fixed protocol state before the computation of message_3,
the plaintext of message_3 is randomized if the corresponding
signature scheme is, but IV_3 and K_3. Allowing to recompute
message_3 leads to a situation where the same key and iv are used
multiple times for distinct message. Such a thing falls outside
of the cryptographicaly authorized usages for an AEAD scheme
[RFC5116, Sec 3.1] and should thus be forbidden.